refactor(version): use of plugin system
