+++ .gitignore
@@ -13,3 +13,5 @@ config.json
 */._*
 */*/._*
 coverage.*
+.eslintrc
+.editorconfig
