+++ lib/index.js
@@ -2,7 +2,7 @@
 
 var Hapi = require('hapi');
 var Hoek = require('hoek');
-var Package = require('../package.json');
+var Version = require('./version');
 
 
 // Declare internals
@@ -14,24 +14,16 @@ internals.init = function () {
 
     var server = new Hapi.Server();
     server.connection({ port: 8000 });
-
-    server.route({
-        method: 'GET',
-        path: '/version',
-        config: {
-            description: 'Returns the version of the server',
-            handler: function (request, reply) {
-
-                return reply({ version: Package.version });
-            }
-        }
-    });
-
-    server.start(function (err) {
+    server.register(Version, function(err) {
 
         Hoek.assert(!err, err);
-        console.log('Server started at: ' + server.info.uri);
+        server.start(function (err) {
+
+            Hoek.assert(!err, err);
+            console.log('Server started at: ' + server.info.uri);
+        });
     });
+
 };
 
 internals.init();
